+++ src/pages/products/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useCartItem } from "./useCartItem";
